refactor: normalize observation and reward statistics in core
