+++ tests/test_checkpoint_state.py
@@ -19,10 +19,10 @@ import pytest
 
 from experiments.checkpoint_store import FileCheckpointStore
 from jax_baselines.core.checkpoint import make_checkpoint_scaffold, snapshot_pytree
+from jax_baselines.core.normalization import RewardNormalizer, RunningMeanStd
 from jax_baselines.CrossQ.crossq import CrossQ
 from jax_baselines.DDPG.ddpg import DDPG
 from jax_baselines.DQN.base_class import Q_Network_Family
-from jax_baselines.math.statistics import RewardNormalizer, RunningMeanStd
 from jax_baselines.SAC.sac import SAC
 from jax_baselines.TD3.td3 import TD3
 from jax_baselines.TD7.td7 import TD7
