+++ tests/test_common_split_contract.py
@@ -27,6 +27,7 @@ EXPECTED_CORE_MODULES = (
     "jax_baselines.core.epoch_buffer",
     "jax_baselines.core.eval",
     "jax_baselines.core.hparams",
+    "jax_baselines.core.normalization",
     "jax_baselines.core.replay_protocol",
     "jax_baselines.core.rollout",
     "jax_baselines.core.rollout_stats",
@@ -73,13 +74,13 @@ def test_no_python_imports_pin_deleted_common_paths():
 def test_representative_symbols_resolve_from_new_homes():
     from jax_baselines.core.checkpoint import CheckpointController
     from jax_baselines.core.env_protocols import EnvInfo
+    from jax_baselines.core.normalization import RunningMeanStd
     from jax_baselines.core.replay_protocol import ReplayBufferFactory
     from jax_baselines.core.runtime_adapters import NoOpLogger
     from jax_baselines.core.training_session import TrainingSession
     from jax_baselines.math.distributional import HLGaussTransform
     from jax_baselines.math.losses import QuantileHuberLosses
     from jax_baselines.math.returns import get_vtrace
-    from jax_baselines.math.statistics import RunningMeanStd
 
     assert CheckpointController
     assert EnvInfo
